functions/6-accounting-engine/requirements.txt
firebase-functions
firebase-admin
google-cloud-firestore
google-cloud-pubsub
google-cloud-bigquery
pandas
